+++ caffe/cppbuild.sh
@@ -43,23 +43,23 @@ esac
 
 GLOG=0.3.4
 GFLAGS=2.1.2
-PROTO=2.6.1
-LEVELDB=1.18
+PROTO=3.1.0
+LEVELDB=1.19
 SNAPPY=1.1.3
 LMDB=0.9.18
-BOOST=1_61_0
+BOOST=1_62_0
 HDF5=1.8.17
-OPENBLAS=0.2.18
+OPENBLAS=0.2.19
 CAFFE_VERSION=master
 
 download https://github.com/google/glog/archive/v$GLOG.tar.gz glog-$GLOG.tar.gz
 download https://github.com/gflags/gflags/archive/v$GFLAGS.tar.gz gflags-$GFLAGS.tar.gz
-download https://github.com/google/protobuf/releases/download/v$PROTO/protobuf-$PROTO.tar.gz protobuf-$PROTO.tar.gz
+download https://github.com/google/protobuf/releases/download/v$PROTO/protobuf-cpp-$PROTO.tar.gz protobuf-$PROTO.tar.gz
 download https://github.com/google/leveldb/archive/v$LEVELDB.tar.gz leveldb-$LEVELDB.tar.gz
 download https://github.com/google/snappy/releases/download/$SNAPPY/snappy-$SNAPPY.tar.gz snappy-$SNAPPY.tar.gz
 download https://github.com/LMDB/lmdb/archive/LMDB_$LMDB.tar.gz lmdb-LMDB_$LMDB.tar.gz
 download http://downloads.sourceforge.net/project/boost/boost/${BOOST//_/.}/boost_$BOOST.tar.gz boost_$BOOST.tar.gz
-download http://www.hdfgroup.org/ftp/HDF5/releases/hdf5-$HDF5/src/hdf5-$HDF5.tar.bz2 hdf5-$HDF5.tar.bz2
+download http://support.hdfgroup.org/ftp/HDF5/releases/hdf5-$HDF5/src/hdf5-$HDF5.tar.bz2 hdf5-$HDF5.tar.bz2
 download https://github.com/xianyi/OpenBLAS/archive/v$OPENBLAS.tar.gz OpenBLAS-$OPENBLAS.tar.gz
 download https://github.com/BVLC/caffe/archive/$CAFFE_VERSION.tar.gz caffe-$CAFFE_VERSION.tar.gz
 
@@ -71,16 +71,16 @@ mkdir -p include lib bin
 OPENCV_PATH="$INSTALL_PATH/../../../opencv/cppbuild/$PLATFORM/"
 
 echo "Decompressing archives..."
-tar --totals -xzf ../glog-$GLOG.tar.gz || true
-tar --totals -xzf ../gflags-$GFLAGS.tar.gz
-tar --totals -xzf ../protobuf-$PROTO.tar.gz
-tar --totals -xzf ../leveldb-$LEVELDB.tar.gz
-tar --totals -xzf ../snappy-$SNAPPY.tar.gz
-tar --totals -xzf ../lmdb-LMDB_$LMDB.tar.gz
-tar --totals -xzf ../boost_$BOOST.tar.gz
-tar --totals -xjf ../hdf5-$HDF5.tar.bz2
-tar --totals -xzf ../OpenBLAS-$OPENBLAS.tar.gz
-tar --totals -xzf ../caffe-$CAFFE_VERSION.tar.gz
+tar --totals -xf ../glog-$GLOG.tar.gz || true
+tar --totals -xf ../gflags-$GFLAGS.tar.gz
+tar --totals -xf ../protobuf-$PROTO.tar.gz
+tar --totals -xf ../leveldb-$LEVELDB.tar.gz
+tar --totals -xf ../snappy-$SNAPPY.tar.gz
+tar --totals -xf ../lmdb-LMDB_$LMDB.tar.gz
+tar --totals -xf ../boost_$BOOST.tar.gz
+tar --totals -xf ../hdf5-$HDF5.tar.bz2
+tar --totals -xf ../OpenBLAS-$OPENBLAS.tar.gz
+tar --totals -xf ../caffe-$CAFFE_VERSION.tar.gz
 
 export CFLAGS="-fPIC"
 export CXXFLAGS="-fPIC"
@@ -107,7 +107,7 @@ cd ..
 
 cd leveldb-$LEVELDB
 make -j $MAKEJ
-cp -a libleveldb.a "$INSTALL_PATH/lib"
+cp -a out-static/libleveldb.a "$INSTALL_PATH/lib"
 cp -a include/leveldb "$INSTALL_PATH/include/"
 cd ..